requirements.txt
click==8.1.3
Flask==2.0.0
Flask-Classful==0.14.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==2.2.0
MarkupSafe==2.1.1
p2pnetwork==1.2
pycryptodome==3.15.0
Werkzeug==2.0.0
zipp==3.8.1
